+++ CHANGELOG.md
@@ -3,6 +3,20 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [Unreleased]
+### Added
+- Added unit test for whitespace handling.
+- Added optional argument `$trim = true` to `$node->text()`
+- Added attribute value normalization
+  - https://www.w3.org/TR/html/syntax.html#attribute-values
+  - https://www.w3.org/TR/xml/#AVNormalize
+### Changed
+- `$node->text()` now supports all block and inline level elements.
+### Deprecated
+### Removed
+### Fixed
+### Security
+
 ## [1.9] - 2019-05-30
 ### Added
 - Added unit test for bug reports